+++ src/termout.c
@@ -289,7 +289,7 @@ attr_rect(cattrflags add, cattrflags sub, cattrflags xor, short y0, short x0, sh
     termline * l = term.lines[y];
     int xl = x0;
     int xr = x1;
-    if (!term.attr_rect) {
+    if (term.attr_rect < 2) {
       if (y != y0)
         xl = term.marg_left;
       if (y != y1)
@@ -3346,8 +3346,10 @@ do_csi(uchar c)
                 term.csi_argv[2] ?: urows, term.csi_argv[3] ?: ucols);
     when CPAIR('*', 'x'):  /* DECSACE: VT420 Select Attribute Change Extent */
       switch (arg0) {
-        when 2: term.attr_rect = true;
-        when 0 or 1: term.attr_rect = false;
+        // use original DECSACE values rather than effective bool value,
+        // so we can respond properly to DECRQSS like xterm
+        when 2: term.attr_rect = 2;
+        when 0 or 1 /*or 2*/: term.attr_rect = arg0;
       }
     when CPAIR('$', 'r')  /* DECCARA: VT420 Change Attributes in Area */
       or CPAIR('$', 't'): {  /* DECRARA: VT420 Reverse Attributes in Area */
@@ -3948,6 +3950,8 @@ do_dcs(void)
         child_printf("\eP1$r%ut\e\\", term.rows);
       } else if (!strcmp(s, "$|")) {  // DECSCPP (columns)
         child_printf("\eP1$r%u$|\e\\", term.cols);
+      } else if (!strcmp(s, "*x")) {  // DECSACE (attribute change extent)
+        child_printf("\eP1$r%u*x\e\\", term.attr_rect);
       } else if (!strcmp(s, "*|")) {  // DECSNLS (lines)
         child_printf("\eP1$r%u*|\e\\", term.rows);
       } else if (!strcmp(s, "$~")) {  // DECSSDT (status line type)